requirements.txt
numpy==1.22.4
pandas==1.4.2
psycopg2==2.9.3
python-dateutil==2.8.2
python-rapidjson==1.6
pytz==2022.1
six==1.16.0
